works for mixed precision and 1-gpu training for input 4, forcast 18

+++ dgmr/dgmr.py
@@ -115,13 +115,17 @@ class DGMR(pl.LightningModule, NowcastingModelHubMixin):
 
     def forward(self, x):
         x = self.generator(x)
+        
         return x
 
     def training_step(self, batch, batch_idx):
-        # print(f"training batch: {batch_idx}, {type(batch)}")
         images, future_images = batch
-        images = images.float()
-        future_images = future_images.float()
+        # images = images.float()
+        # future_images = future_images.float()
+        
+        # images = images.to(torch.float16)
+        # future_images = future_images.to(torch.float16)
+        
         self.global_iteration += 1
         g_opt, d_opt = self.optimizers()
         ##########################
@@ -195,11 +199,7 @@ class DGMR(pl.LightningModule, NowcastingModelHubMixin):
             )
 
     def validation_step(self, batch, batch_idx):
-        # print(f"valid batch: {batch_idx}, {type(batch[0])}")
         images, future_images = batch
-        # print(f"images: {images.shape}, future_images: {future_images.shape}")
-        images = images.float()
-        future_images = future_images.float()
         ##########################
         # Optimize Discriminator #
         ##########################
